Seeder更新

+++ database/migrations/2023_12_20_093528_create_areas_finely_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('area_id')->constrained('areas')->onDelete('cascade');
             $table->string('name',100);
-            $table->integer('latitude'); //緯度
-            $table->integer('longitude'); //経度
+            $table->decimal('latitude'); //緯度
+            $table->decimal('longitude'); //経度
             $table->string('path');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent();
